+++ dune/dpg/integralterm_ru_impl.hh
@@ -41,7 +41,7 @@ inline static void interiorImpl(const LhsLocalView& lhsLocalView,
       ::Quadrature(element, quadratureOrder);
 
   const auto referenceGridView =
-      lhsLocalView.tree().refinedReferenceElement().leafGridView();
+      lhsLocalView.tree().refinedReferenceElementGridView();
 
   const unsigned int subElementStride =
       (is_DGRefinedFiniteElement<LhsSpace>::value) ?
@@ -143,7 +143,7 @@ faceImpl(const LhsLocalView& lhsLocalView,
   const unsigned int nRhs(rhsLocalFiniteElement.localBasis().size());
 
   const auto referenceGridView =
-      lhsLocalView.tree().refinedReferenceElement().leafGridView();
+      lhsLocalView.tree().refinedReferenceElementGridView();
 
   const unsigned int subElementStride =
       (is_DGRefinedFiniteElement<LhsSpace>::value) ?